+++ Gopkg.toml
@@ -26,12 +26,16 @@
 
 
 [[constraint]]
-  branch = "master"
   name = "github.com/dgrijalva/jwt-go"
+  version = "3.2.0"
 
 [[constraint]]
   name = "github.com/gin-gonic/gin"
   version = "1.3.0"
+
+[[constraint]]
+  name = "github.com/jinzhu/gorm"
+  version = "1.9.2"
 
 [[constraint]]
   name = "github.com/lib/pq"
@@ -42,8 +46,12 @@
   version = "1.2.2"
 
 [[constraint]]
+  branch = "master"
+  name = "golang.org/x/crypto"
+
+[[constraint]]
   name = "gopkg.in/go-playground/validator.v8"
-  version = "8.18.2"
+  version = "8.18.1"
 
 [[constraint]]
   name = "gopkg.in/gomail.v2"
@@ -52,3 +60,7 @@
 [prune]
   go-tests = true
   unused-packages = true
+
+[[constraint]]
+  name = "github.com/golang-migrate/migrate"
+  version = "3.5.4"
