+++ app/common/database.go
@@ -2,10 +2,13 @@ package common
 
 import (
 	"books-backend/app"
+	"database/sql"
 	"fmt"
+	"github.com/golang-migrate/migrate"
+	"github.com/golang-migrate/migrate/database/postgres"
+	_ "github.com/golang-migrate/migrate/source/file"
 	"github.com/jinzhu/gorm"
-	_ "github.com/jinzhu/gorm/dialects/sqlite"
-	"os"
+	_ "github.com/lib/pq"
 )
 
 type Database struct {
@@ -41,27 +44,64 @@ func GetDB() *gorm.DB {
 }
 
 func TestDBInit() *gorm.DB {
-	//out, err := exec.Command("/bin/sh", "../../db_init.sh testbooks").Output()
+	test_db_name := "test_books"
+	connectionString := fmt.Sprintf(
+		"port=%s user=%s password=%s dbname=%s sslmode=disable",
+		app.DB_PORT,
+		app.DB_USER,
+		app.DB_PASSWORD,
+		"postgres")
+	fmt.Println(connectionString)
+	db, err := sql.Open("postgres", connectionString)
+	if err != nil {
+		panic(err)
+	}
+	_, err = db.Exec("CREATE DATABASE test_books")
+	db.Close()
+	connectionString = fmt.Sprintf(
+		"port=%s user=%s password=%s dbname=%s sslmode=disable",
+		app.DB_PORT,
+		app.DB_USER,
+		app.DB_PASSWORD,
+		test_db_name)
+	db, err = sql.Open("postgres", connectionString)
+	defer db.Close()
+	if err != nil {
+		panic(err)
+	}
+	driver, err := postgres.WithInstance(db, &postgres.Config{})
+	if err != nil {
+		panic(err)
+	}
+	//dir, err := os.Getwd()
 	//if err != nil {
-	//	log.Fatal(err)
+	//	panic(err)
 	//}
-	//fmt.Printf("output is %s\n", out)
-	//test_db, err := gorm.Open("postgres", "host=localhost port=5432 user=testbooks dbname=testbooks password=testbooks sslmode=disable")
-	test_db, err := gorm.Open("sqlite3", "./../gorm_test.db")
+	m, err := migrate.NewWithDatabaseInstance(
+		`file://../../migrations`,
+		"postgres", driver)
 	if err != nil {
-		fmt.Println("db err: ", err)
+		panic(err)
 	}
-	if err != nil {
-		fmt.Println("db err: ", err)
-	}
-	test_db.DB().SetMaxIdleConns(3)
-	test_db.LogMode(true)
+	m.Up()
+	test_db, err := gorm.Open("postgres", connectionString)
 	DB = test_db
 	return DB
 }
 
-func TestDBFree(test_db *gorm.DB) error {
+func TestDBFree() error {
 	//test_db.Close()
-	err := os.Remove("./../gorm_test.db")
+	connectionString := fmt.Sprintf(
+		"port=%s user=%s password=%s dbname=%s sslmode=disable",
+		app.DB_PORT,
+		app.DB_USER,
+		app.DB_PASSWORD,
+		"postgres")
+	db, err := sql.Open("postgres", connectionString)
+	fmt.Println(err)
+	db.Exec("SELECT pg_terminate_backend(pg_stat_activity.pid) " +
+		"FROM pg_stat_activity " +
+		"WHERE pg_stat_activity.datname = 'test_books' AND pid <> pg_backend_pid();")
+	_, err = db.Exec("DROP DATABASE test_books")
 	return err
 }
